feat(modularize): separate out server
